Client/src/requirements.txt
requests
pycryptodome
urllib
